✨ Prepare protocol for system calls
